tweak: don't native-compile the profile loader
...to avoid repeatedly recompiling it.

When Doom's early-init.el runs the profile loader,
native-comp-eln-load-path only has its usual Emacs paths on it. If Emacs
does not find the corresponding .eln file there, it queues it for
deferred compilation.

But if the profile loader is only used to load Doom, then by the time
deferred compilation happens Doom has updated native-comp-eln-load-path,
so deferred compilation writes to Doom's cache dir, not the "normal"
Emacs one. Emacs will not find it there on next startup, so this
repeats.

This is harmless (but noticeable through the
`*Async-native-compile-log*` buffer appearing every time Doom starts
up), but needless work at startup that would be nice to avoid.

Disabling native-compilation for this file seems much easier than
getting the .eln file written to the correct location, and it seems
unlikely code benefitting from native-comp would end up in the profile
loader.

(An even easier fix would be to stop byte-compiling the profile loader:
I'm not sure why Doom does this...)